refactor(#119): adopt Waaseyaa Application Directory Convention v1.0
Move Geo/ → Domain/Geo/ (Service/, ValueObject/), rename Ingest/ →
Ingestion/, extract cross-cutting utilities to Support/. Update all
namespaces and references across src/ and tests/.
