+++ src/Ingestion/EntityMapper/WordPartMapper.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Minoo\Ingest\EntityMapper;
+namespace Minoo\Ingestion\EntityMapper;
 
-use Minoo\Ingest\SlugGenerator;
-use Minoo\Ingest\ValueObject\WordPartFields;
+use Minoo\Support\SlugGenerator;
+use Minoo\Ingestion\ValueObject\WordPartFields;
 
 final class WordPartMapper
 {
