+++ src/Ingestion/ValueObject/SpeakerFields.php
@@ -2,15 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Minoo\Ingest\ValueObject;
+namespace Minoo\Ingestion\ValueObject;
 
-final readonly class WordPartFields
+final readonly class SpeakerFields
 {
     public function __construct(
-        public string $form,
-        public string $type,
-        public string $definition,
-        public string $sourceUrl,
+        public string $name,
+        public string $code,
+        public ?string $bio,
         public string $slug,
         public int $status,
         public int $createdAt,
@@ -21,10 +20,9 @@ final readonly class WordPartFields
     public function toArray(): array
     {
         return [
-            'form' => $this->form,
-            'type' => $this->type,
-            'definition' => $this->definition,
-            'source_url' => $this->sourceUrl,
+            'name' => $this->name,
+            'code' => $this->code,
+            'bio' => $this->bio,
             'slug' => $this->slug,
             'status' => $this->status,
             'created_at' => $this->createdAt,
