+++ tests/Minoo/Unit/Geo/LocationServiceTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Minoo\Tests\Unit\Geo;
 
-use Minoo\Geo\LocationService;
+use Minoo\Domain\Geo\Service\LocationService;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
