+++ tests/Minoo/Unit/Geo/VolunteerRankerTest.php
@@ -7,8 +7,8 @@ namespace Minoo\Tests\Unit\Geo;
 use Minoo\Entity\Community;
 use Minoo\Entity\ElderSupportRequest;
 use Minoo\Entity\Volunteer;
-use Minoo\Geo\RankedVolunteer;
-use Minoo\Geo\VolunteerRanker;
+use Minoo\Domain\Geo\ValueObject\RankedVolunteer;
+use Minoo\Domain\Geo\Service\VolunteerRanker;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
