+++ tests/Minoo/Unit/Ingest/EntityMapper/SpeakerMapperTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Minoo\Tests\Unit\Ingest\EntityMapper;
 
-use Minoo\Ingest\EntityMapper\SpeakerMapper;
-use Minoo\Ingest\ValueObject\SpeakerFields;
+use Minoo\Ingestion\EntityMapper\SpeakerMapper;
+use Minoo\Ingestion\ValueObject\SpeakerFields;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
