+++ tests/Minoo/Unit/Ingest/MaterializationContextTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Minoo\Tests\Unit\Ingest;
 
-use Minoo\Ingest\MaterializationContext;
-use Minoo\Ingest\MaterializationResult;
+use Minoo\Ingestion\MaterializationContext;
+use Minoo\Ingestion\MaterializationResult;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
